+++ src/main/java/com/lmax/disruptor/Sequence.java
@@ -43,7 +43,6 @@ public class Sequence extends RhsPadding
         {
             throw new RuntimeException(e);
         }
-
     }
 
     /**